share icon refactor
