+++ src/components/buttons/share-button/index.tsx
@@ -1,0 +1,2 @@
+import ShareButton from "./ShareButton";
+export { ShareButton };
